pyproject.toml
[project]
name = "webtool-platform-backend"
version = "2.0.0"
description = "AI-powered web tool platform backend with FastAPI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.6.3",
    "python-decouple>=3.8",
    "python-dotenv>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]
